# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: BETCHU API
  version: 0.3.0
  description: BETCHU Google 로그인, 연령·정책 동의, 계정 활성화와 안전한 커플 연결 계약
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
paths:
...
  /auth/providers:
    get:
      operationId: getAuthProviders
      security: []
      responses:
        "200":
          description: Configured login providers and registration readiness
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthProviders"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - onboarding
  /auth/login/start:
    post:
      operationId: startLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginStartRequest"
      responses:
        "200":
          description: One-use attempt; retain loginSecret locally and open authorizationUrl
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginStart"
        "400":
          $ref: "#/components/responses/AuthBadRequest"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - onboarding
  /auth/oauth/google/callback:
    get:
      operationId: googleOAuthCallback
      security: []
      description: Server-only Google OAuth redirect. The allowlisted app receives loginId and status SUCCESS or
        FAILED. Only SUCCESS also carries a one-use handoffCode valid for at most 120 seconds. No access
        token, refresh token, or loginSecret is redirected. The app removes callback credentials from the URL
        immediately and supplies both local loginSecret and callback handoffCode to complete login.
      parameters:
        - in: query
          name: state
          required: true
          schema:
            type: string
        - in: query
          name: code
          schema:
            type: string
        - in: query
          name: error
          schema:
            type: string
      responses:
        "302":
          description: Return to the allowlisted app callback
          headers:
            Location:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/AuthBadRequest"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - onboarding
  /auth/login:
    post:
      operationId: completeLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Atomically verify both the local loginSecret and callback-only handoffCode, consume a verified
            attempt, and issue revocable tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "400":
          $ref: "#/components/responses/AuthBadRequest"
        "401":
          $ref: "#/components/responses/AuthUnauthorized"
        "409":
          $ref: "#/components/responses/AuthConflict"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - onboarding
  /auth/refresh:
    post:
      operationId: refreshSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Rotate the refresh token; the previous refresh token becomes unusable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "400":
          $ref: "#/components/responses/AuthBadRequest"
        "401":
          $ref: "#/components/responses/AuthUnauthorized"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - onboarding
  /auth/logout:
    post:
      operationId: logoutSession
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Current session revoked immediately
          including its access and refresh tokens: null
        "401":
          $ref: "#/components/responses/AuthUnauthorized"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - onboarding
  /users/me:
    get:
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current account and current policy consent status; available to limited onboarding sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserMe"
        "401":
          $ref: "#/components/responses/AuthUnauthorized"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - onboarding
  /policy-versions/current:
    get:
      operationId: getCurrentPolicyVersions
      security: []
      parameters:
        - in: query
          name: locale
          schema:
            type: string
            default: ko-KR
            enum:
              - ko-KR
      responses:
        "200":
          description: Ready only when current effective approved TERMS and PRIVACY documents exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentPolicies"
        "400":
          $ref: "#/components/responses/AuthBadRequest"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - onboarding
  /onboarding/age-eligibility:
    post:
      operationId: confirmAgeEligibility
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgeEligibilityRequest"
      responses:
        "200":
          description: Age confirmed; account activates only if all current required policies are accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserMe"
        "204":
          description: Age rejected or declined; pending user and every limited session deleted immediately
        "400":
          $ref: "#/components/responses/AuthBadRequest"
        "401":
          $ref: "#/components/responses/AuthUnauthorized"
        "409":
          $ref: "#/components/responses/AuthConflict"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - onboarding
  /consents/{policyType}:
    put:
      operationId: acceptPolicy
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: policyType
          required: true
          schema:
            $ref: "#/components/schemas/PolicyType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentRequest"
      responses:
        "200":
          description: Idempotent acceptance of the current effective policy version; may activate the account and grant
            signup coins once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserMe"
        "400":
          $ref: "#/components/responses/AuthBadRequest"
        "401":
          $ref: "#/components/responses/AuthUnauthorized"
        "409":
          $ref: "#/components/responses/AuthConflict"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - onboarding
  /couples/me:
    get:
      operationId: getMyCouple
      responses:
        "200":
          description: Current accessible relationship and pending invitation; both null when unconnected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoupleState"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - couples
  /couples/invites:
    post:
      operationId: createCoupleInvite
      parameters:
        - $ref: "#/components/parameters/CoupleIdempotencyKey"
      description: A fresh key rotates the existing invitation. A repeated key returns the original metadata with
        code null, even after expiry or rotation; the raw code is never stored or shown again.
      responses:
        "200":
          description: New invitation or idempotent metadata replay.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedCoupleInvite"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - couples
  /couples/invites/{inviteId}:
    delete:
      operationId: revokeCoupleInvite
      parameters:
        - $ref: "#/components/parameters/CoupleInviteId"
      responses:
        "204":
          description: Revoked; repeat revocation is safe.
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - couples
  /couples/invites/preview:
    post:
      operationId: previewCoupleInvite
      requestBody:
        $ref: "#/components/requestBodies/CoupleCode"
      responses:
        "200":
          description: Only the inviter nickname/profile image; no wallet or other private data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoupleInvitePreview"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - couples
  /couples/join:
    post:
      operationId: joinCoupleInvite
      requestBody:
        $ref: "#/components/requestBodies/CoupleCode"
      responses:
        "200":
          description: Reserves this user as the invitee; does not connect until both users confirm. Repeating this
            claim returns the same pending state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingCoupleInvite"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - couples
  /couples/pending/{inviteId}/confirm:
    post:
      operationId: confirmCoupleInvite
      parameters:
        - $ref: "#/components/parameters/CoupleInviteId"
      responses:
        "200":
          description: This user's confirmation or an idempotent replay; the second confirmation connects atomically.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoupleState"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - couples
  /couples/pending/{inviteId}/reject:
    post:
      operationId: rejectCoupleInvite
      parameters:
        - $ref: "#/components/parameters/CoupleInviteId"
      responses:
        "204":
          description: Rejected and both slots released; repeats are safe.
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - couples
  /couples/me/end:
    post:
      operationId: endMyCouple
      parameters:
        - $ref: "#/components/parameters/CoupleIdempotencyKey"
      responses:
        "200":
          description: Unilateral immediate relationship end. Current increment has no quest/resources to settle.
            Replays never end a later relationship.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoupleEndResult"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - couples
  /couples/me/block:
    post:
      operationId: blockMyPartner
      parameters:
        - $ref: "#/components/parameters/CoupleIdempotencyKey"
      responses:
        "200":
          description: Immediately ends and blocks the current partner; no arbitrary target ID. Replays never block a
            later partner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoupleEndResult"
        default:
          description: Request rejected with a stable BETCHU error code
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      tags:
        - couples
components:
  schemas:
...
    ApiProblem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - errorCode
      properties:
        type:
          type: string
          const: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        errorCode:
          type: string
    AuthProviders:
      type: object
      required:
        - providers
        - registrationAvailable
      properties:
        providers:
          type: array
          items:
            type: object
            required:
              - provider
              - displayName
              - enabled
            properties:
              provider:
                type: string
                const: GOOGLE
              displayName:
                type: string
              enabled:
                type: boolean
        registrationAvailable:
          type: boolean
    LoginStartRequest:
      type: object
      additionalProperties: false
      required:
        - provider
        - redirectUri
      properties:
        provider:
          type: string
          const: GOOGLE
        redirectUri:
          type: string
          maxLength: 1000
    LoginStart:
      type: object
      required:
        - loginId
        - loginSecret
        - authorizationUrl
        - expiresAt
      properties:
        loginId:
          type: string
          format: uuid
        loginSecret:
          type: string
        authorizationUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - loginId
        - loginSecret
        - handoffCode
      properties:
        loginId:
          type: string
          format: uuid
        loginSecret:
          type: string
          minLength: 1
          maxLength: 128
        handoffCode:
          type: string
          minLength: 1
          maxLength: 128
          description: One-use callback proof; valid for at most 120 seconds and stored only as a hash on the server
    RefreshRequest:
      type: object
      additionalProperties: false
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          minLength: 1
          maxLength: 128
    AuthTokens:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          const: 300
        user:
          $ref: "#/components/schemas/UserMe"
    UserMe:
      type: object
      required:
        - id
        - nickname
        - status
        - ageEligible
        - requiredPolicyVersionIds
        - missingPolicyVersionIds
        - evidenceConsent
        - availableCoins
        - lockedCoins
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
          minLength: 1
          maxLength: 30
        status:
          type: string
          enum:
            - PENDING_ELIGIBILITY
            - ACTIVE
            - DELETION_PENDING
            - DELETED
        ageEligible:
          type: boolean
        requiredPolicyVersionIds:
          type: array
          items:
            type: string
            format: uuid
        missingPolicyVersionIds:
          type: array
          items:
            type: string
            format: uuid
        evidenceConsent:
          type: boolean
        availableCoins:
          type:
            - integer
            - "null"
          minimum: 0
        lockedCoins:
          type:
            - integer
            - "null"
          minimum: 0
    PolicyType:
      type: string
      enum:
        - TERMS
        - PRIVACY
        - EVIDENCE_OPTIONAL
    PolicyVersion:
      type: object
      required:
        - id
        - policyType
        - version
        - locale
        - required
        - documentUrl
        - effectiveAt
      properties:
        id:
          type: string
          format: uuid
        policyType:
          $ref: "#/components/schemas/PolicyType"
        version:
          type: string
        locale:
          type: string
          const: ko-KR
        required:
          type: boolean
        documentUrl:
          type: string
          format: uri
        effectiveAt:
          type: string
          format: date-time
    CurrentPolicies:
      type: object
      required:
        - ready
        - policies
      properties:
        ready:
          type: boolean
        policies:
          type: array
          items:
            $ref: "#/components/schemas/PolicyVersion"
    AgeEligibilityRequest:
      type: object
      additionalProperties: false
      required:
        - eligible
      properties:
        eligible:
          type:
            - boolean
            - "null"
    ConsentRequest:
      type: object
      additionalProperties: false
      required:
        - policyVersionId
        - locale
      properties:
        policyVersionId:
          type: string
          format: uuid
        locale:
          type: string
          const: ko-KR
    CouplePublicProfile:
      type: object
      additionalProperties: false
      required:
        - id
        - nickname
        - profileImage
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        profileImage:
          type:
            - string
            - "null"
    CurrentCouple:
      type: object
      additionalProperties: false
      required:
        - id
        - connectedAt
        - partner
      properties:
        id:
          type: string
          format: uuid
        connectedAt:
          type: string
          format: date-time
        partner:
          $ref: "#/components/schemas/CouplePublicProfile"
    PendingCoupleInvite:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - role
        - expiresAt
        - partner
        - myConfirmed
        - partnerConfirmed
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - ACTIVE
            - PENDING_CONFIRMATION
        role:
          type: string
          enum:
            - INVITER
            - INVITEE
        expiresAt:
          type: string
          format: date-time
        partner:
          oneOf:
            - $ref: "#/components/schemas/CouplePublicProfile"
            - type: "null"
        myConfirmed:
          type: boolean
        partnerConfirmed:
          type: boolean
    CoupleState:
      type: object
      additionalProperties: false
      required:
        - couple
        - pendingInvite
      properties:
        couple:
          oneOf:
            - $ref: "#/components/schemas/CurrentCouple"
            - type: "null"
        pendingInvite:
          oneOf:
            - $ref: "#/components/schemas/PendingCoupleInvite"
            - type: "null"
    CreatedCoupleInvite:
      type: object
      additionalProperties: false
      required:
        - inviteId
        - code
        - expiresAt
      properties:
        inviteId:
          type: string
          format: uuid
        code:
          type:
            - string
            - "null"
          description: 16 random characters in four hyphen-separated groups; null on replay.
        expiresAt:
          type: string
          format: date-time
    CoupleInvitePreview:
      type: object
      additionalProperties: false
      required:
        - inviteId
        - inviter
        - expiresAt
      properties:
        inviteId:
          type: string
          format: uuid
        inviter:
          $ref: "#/components/schemas/CouplePublicProfile"
        expiresAt:
          type: string
          format: date-time
    CoupleEndResult:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          const: COMPLETED
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    AuthBadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ApiProblem"
    AuthUnauthorized:
      description: Invalid, expired, consumed or revoked credentials
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ApiProblem"
    AuthConflict:
      description: Login pending, obsolete policy version, or account state conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ApiProblem"
    AuthUnavailable:
      description: Required external authentication settings are unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ApiProblem"
  parameters:
    CoupleIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
    CoupleInviteId:
      name: inviteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  requestBodies:
    CoupleCode:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required:
              - code
            properties:
              code:
                type: string
                maxLength: 64
security:
  - BearerAuth: []